% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.let.R
\name{add.let}
\alias{add.let}
\title{Add Letter}
\usage{
add.let(word)
}
\arguments{
\item{word}{Word to swap letters}
}
\value{
Returns a vector with all possible added letter combinations
}
\description{
A wrapper function for spell-checking
(adds additional nearby letters at each letter of the word)
}
\examples{

add.let("bombay")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
