% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.database.R
\docType{data}
\name{animals.database}
\alias{animals.database}
\title{Animals Database}
\format{animals.database (vector, length = 848)}
\usage{
data(animals.database)
}
\description{
A database of possible animals responses (\emph{n} = 848)
}
\details{
To add additional animals to the database, please email me at
<alexpaulchristensen@gmail.com>
}
\examples{
data("animals.database")
}
\keyword{datasets}
