% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.misnomer.R
\docType{data}
\name{animals.misnomer}
\alias{animals.misnomer}
\title{Animals Misnomer Database}
\format{animals.misnomer (list, length = 99)}
\usage{
data(animals.misnomer)
}
\description{
A database of possible animals misnomers
}
\details{
To add additional animals misnomers to the database, please email me at
<alexpaulchristensen@gmail.com>
}
\examples{
data("animals.misnomer")
}
\keyword{datasets}
