% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.guess.R
\name{best.guess}
\alias{best.guess}
\title{Best guess function}
\usage{
best.guess(vec, database)
}
\arguments{
\item{vec}{Vector of potential options}

\item{database}{The database to search for best guesses in}
}
\value{
The best guess(es) of the word
}
\description{
A wrapper function for best guess spelling
}
\examples{

best.guess("bomba", animals.database)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
