% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chn.let.R
\name{chn.let}
\alias{chn.let}
\title{Change Letter}
\usage{
chn.let(word)
}
\arguments{
\item{word}{Word to swap letters}
}
\value{
Returns a vector with each letter swapped for another
}
\description{
A wrapper function for spell-checking
(swaps each letter of the word with another nearby letter)
}
\examples{

chn.let("bombae")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
