% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.responses.R
\name{combine.responses}
\alias{combine.responses}
\title{Combine Words Wrapper}
\usage{
combine.responses(vec, dictionary)
}
\arguments{
\item{vec}{Vector.
A vector with words to potentially be combined}

\item{dictionary}{A dictionary to look for (combined) word in.
See \code{\link{SemNetDictionaries}}}
}
\value{
A vector with responses combined based on dictionary entries
}
\description{
A wrapper function to combine words that are found in dictionary
(e.g., "star" "fish" --> "starfish")
}
\examples{
# Convert "star fish" to "starfish"
combine.responses("star fish", SemNetDictionaries::animals.dictionary)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
