% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.chn.R
\name{corr.chn}
\alias{corr.chn}
\title{Correct Changes from \code{\link[SemNetCleaner]{textcleaner}}}
\usage{
corr.chn(txt.cln.obj, resp, corr)
}
\arguments{
\item{txt.cln.obj}{A \code{\link[SemNetCleaner]{textcleaner}} object}

\item{resp}{A vector of response(s) to change.
See the object \code{spellcheck$unique} in
\code{\link[SemNetCleaner]{textcleaner}} output}

\item{corr}{How should the response(s) be corrected?
\code{1} = KEEP the original response.
\code{0} = REMOVE the response entirely.
These must correspond to the input for the \code{resp}
argument}
}
\value{
This function returns a list containing the
following \code{\link[SemNetCleaner]{textcleaner}} objects, which
have been changed with the provided changes:

\item{binary}{A matrix of responses where each row represents a participant
and each column represents a unique response. A response that a participant has provided is a '1'
and a response that a participant has not provided is a '0'}

\item{resposnes}{A response matrix that has been spell checked and de-pluralized with duplicates removed}

\item{spellcheck}{A list containing two objects: full and unique. \strong{full} contains
all responses regardless of spellcheck changes and \strong{unique} contains only responses that were
changed during the spell check}

\item{removed}{A list containing two objects: rows and ids.
\strong{rows} identifies removed participants by their row location in the original data file
and \strong{ids} identifies removed participants by their ID}

\item{partChanges}{A list where each participant is an object with their
responses that have been changed. Participants are identified by their ID.
This can be used to replicate the cleaning process and to keep track of changes more generaly.
Participants with \strong{NA} did not have any changes from the original data
and participants with NULL were removed due to missing data (see \emph{removed$ids})}
}
\description{
Allows from correction of changes made by \code{\link[SemNetCleaner]{textcleaner}}.
Some changes may have been made by accident, some changes may have been made
by the automated cleaning, while others may simply just need to be removed.
This function will correct any changes made in a cleaned \code{\link[SemNetCleaner]{textcleaner}}
object.
}
\examples{

\donttest{
tc.obj <- textcleaner(trial, partBY = "col")

rmat <- corr.chn (tc.obj, "rat", 1)
}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
