% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize}
\alias{finalize}
\title{Finalize Response Matrix}
\usage{
finalize(rmat, minCase = 2)
}
\arguments{
\item{rmat}{Binary matrix.
A \link[SemNetCleaner]{textcleaner} filtered response matrix}

\item{minCase}{Numeric.
Minumum number of cases to produce a response}
}
\value{
A binary response matrix with responses
given by at least \code{minCase} people
}
\description{
Finalizes the response matrix by keeping
responses that are given by two or more people
}
\examples{
\dontrun{

convmat <- autoConverge(rmat)
}

finalRmat <- finalize(convmat)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
