% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.letter.R
\name{is.letter}
\alias{is.letter}
\title{Checks If A Character Is A Letter}
\usage{
is.letter(letter)
}
\arguments{
\item{letter}{A single character}
}
\value{
A \code{TRUE} or \code{FALSE} value for whether
the character entered is a letter
}
\description{
A wrapper function designed to determine whether
a single character is a letter
}
\examples{
#TRUE
is.letter("r")

#FALSE
is.letter("5")

#FALSE
is.letter("~")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
