% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead.word.R
\docType{data}
\name{lead.word}
\alias{lead.word}
\title{Lead Word Example}
\format{lead.word (vector, length = 1)}
\usage{
data(lead.word)
}
\description{
An example word demonstrating that \code{\link{trimws}} does
not remove leading space for a word
}
\examples{
data("lead.word")
}
\keyword{datasets}
