% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter.freq.R
\docType{data}
\name{letter.freq}
\alias{letter.freq}
\title{Letter Frequencies Based on 40,000 Words}
\format{
letter.freq (26-element numeric vector)
}
\usage{
data(letter.freq)
}
\description{
A vector corresponding the frequency of letters across 40,000 words.
Retrieved from: http://pi.math.cornell.edu/~mec/2003-2004/cryptography/subs/frequencies.html
}
\examples{
data("letter.freq")
}
\keyword{datasets}
