% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misnomer.R
\name{misnomer}
\alias{misnomer}
\title{Misnomer Fucntion}
\usage{
misnomer(word, misnom)
}
\arguments{
\item{word}{Word to check for misnomer}

\item{misnom}{Misnomer database}
}
\value{
If \code{word} matches a misnomer, then the appropriate word is returned.
If \code{word} does not match a misnomer, then the \code{word} is returned
}
\description{
A wrapper function for spell-checking
(idnetifies misnomers for a word)
}
\examples{
misnomer("possum", animals.misnomer)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
