% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.let.R
\name{miss.let}
\alias{miss.let}
\title{Missing Letter}
\usage{
miss.let(word)
}
\arguments{
\item{word}{Word to remove letters}
}
\value{
Returns a vector with each letter from the word missing
}
\description{
A wrapper function for spell-checking
(removes each letter from the word)
}
\examples{

miss.let("bombay")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
