% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moniker.R
\name{moniker}
\alias{moniker}
\title{Moniker Function}
\usage{
moniker(word, misnom)
}
\arguments{
\item{word}{Word to check for moniker}

\item{misnom}{A list of monikers.
See \code{\link[SemNetDictionaries]{dictionaries}} for options}
}
\value{
If \code{word} matches a moniker, then the appropriate word is returned.
If \code{word} does not match a moniker, then the \code{word} is returned
}
\description{
A wrapper function for spell-checking
(identifies monikers for a word)
}
\examples{
moniker("possum", SemNetDictionaries::animals.moniker)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
