% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Max-min normalization}
\usage{
normalize(vec)
}
\arguments{
\item{vec}{Vector of potential options}
}
\value{
A vector of normalized values ranging from 0 to 1
}
\description{
A wrapper function for best guess spelling
}
\examples{
#create a numeric vector list
num <- runif(n = 100, min = 0, max = 100)

#normalize the number vector
normalize(num)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
