% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.preprocess.R
\docType{data}
\name{open.preprocess}
\alias{open.preprocess}
\title{Preprocessed \code{\link[SemNetCleaner]{textcleaner}} Object (Openness and Verbal Fluency)}
\format{
open.preprocess (list, length = 4)
}
\usage{
data(open.preprocess)
}
\description{
Preprocessed \code{\link[SemNetCleaner]{textcleaner}} object for the Animals verbal fluency data (\emph{n} = 516)
from Christensen and Kenett (2020).
}
\examples{
data("open.preprocess")
}
\references{
Christensen, A. P., & Kenett, Y. N. (2020).
Semantic network analysis (SemNA): A tutorial on preprocessing, estimating, and analyzing semantic networks.
\emph{PsyArxiv}.
}
\keyword{datasets}
