% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.lead.space.R
\name{rm.lead.space}
\alias{rm.lead.space}
\title{Removes Leading Spaces}
\usage{
rm.lead.space(word)
}
\arguments{
\item{word}{Character (vector). A word that has leading
spaces that cannot be removed by \code{\link{trimws}}}
}
\value{
Word without leading spaces
}
\description{
Removes leading spaces that are not caught by \code{\link{trimws}}
}
\examples{
# 'trimws' should remove lead space but doesn't
trimws(lead.word)

# 'rm.lead.space' does
rm.lead.space(lead.word)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
