% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemNetCleaner.R
\docType{package}
\name{SemNetCleaner-package}
\alias{SemNetCleaner}
\alias{SemNetCleaner-package}
\title{SemNetCleaner--package}
\description{
This package is mainly a wrapper package for functions called in SemNetToolbox.
It implements several functions that automatize the cleaning, removal of plurals and
continuous strings, making the data binary, converging, and finalizing of linguistic
data for semantic network analysis. It also includes databases to make the cleaning process
more accurate, efficient, and reproducible.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
