% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell.check.database.R
\name{spell.check.database}
\alias{spell.check.database}
\title{Spelling-check for Dababases}
\usage{
spell.check.database(data, database)
}
\arguments{
\item{data}{Input data to spell check}

\item{database}{The database to spell check from}
}
\value{
Returns a vector of cleaned response
}
\description{
A wrapper function for spell-checking text databases in \code{\link{SemNetCleaner}}
(combines all spell-checking wrapper functions)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
