% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell.opt.R
\name{spell.opt}
\alias{spell.opt}
\title{All Possible Spelling Options}
\usage{
spell.opt(word)
}
\arguments{
\item{word}{Word to check spelling options}
}
\value{
Returns a vector of all possible spelling combinations from
several spell-checking options
}
\description{
A wrapper function for spell-checking
(combines
\link[SemNetCleaner]{singularize},
\link[SemNetCleaner]{miss.let},
\link[SemNetCleaner]{chn.let}, and
\link[SemNetCleaner]{add.let} 
spell-checking wrapper functions)
}
\examples{
#bombay
spell.opt("bomba")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
