% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitstr.check.R
\name{splitstr.check}
\alias{splitstr.check}
\title{Split String Check}
\usage{
splitstr.check(split, base)
}
\arguments{
\item{split}{A split string to check}

\item{base}{Database to check}
}
\value{
Returns the string as is or with the selected responses merged
}
\description{
A wrapper function for spell-checking
(ensures next word does not belong to the previous)
}
\examples{
#create long word vector
vec <- "bombay opossum guinea pig horse cow"

#split vector
split <- strsplit(vec, " ")[[1]]

\donttest{
splitstr.check(split, animals.database)
}

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
