% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate.multi.R
\name{equate.multi}
\alias{equate.multi}
\title{Equate Multiple Groups}
\usage{
equate.multi(...)
}
\arguments{
\item{...}{Datasets to be equated}
}
\value{
This function returns a list containing equated datasets
in the order they were input. They are labeled dataset (1, ..., n)
}
\description{
An automated cleaning function for matching groups' responses.
Handles multiple datasets
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
