% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize}
\alias{finalize}
\title{Finalize Response Matrix}
\usage{
finalize(rmat, minCase = 2)
}
\arguments{
\item{rmat}{Binary matrix.
A \link[SemNetCleaner]{textcleaner} filtered response matrix}

\item{minCase}{Numeric.
Minimum number of cases to produce a response}
}
\value{
A binary response matrix with responses
given by at least \code{minCase} people
}
\description{
Finalizes the response matrix by keeping
responses that are given by a certain number of people
}
\examples{
# Toy example
raw <- open.animals[c(1:10),-c(1:3)]

# Clean and prepocess data
clean <- textcleaner(raw, partBY = "row", dictionary = "animals")

# Obtain binary data
bin <- clean$binary

# Finalize mat1
mat1 <- finalize(bin)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
