% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.animals.R
\docType{data}
\name{open.animals}
\alias{open.animals}
\title{Openness and Verbal Fluency}
\format{
open.animals (matrix 516 x 38)
}
\usage{
data(open.animals)
}
\description{
Raw Animals verbal fluency data (\emph{n} = 516) from Christensen et al. (2018).
}
\details{
First column is a grouping variable (\code{"Group"}) with \code{1} corresponding
to low openness to experience and \code{2} to high openness to experience

Second column is the latent variable of openness to experience with Intellect items removed
(see Christensen et al., 2018 for more details).

Third column is the ID variable for each participant.

Columns 4-38 are raw fluency data.
}
\examples{
data("open.animals")
}
\references{
Christensen, A. P., Kenett, Y. N., Cotter, K. N., Beaty, R. E., & Silvia, P. J. (2018).
Remotely close associations: Openness to experience and semantic memory structure.
\emph{European Journal of Personality}, \emph{32}, 480-492.
doi:\href{https://doi.org/10.1002/per.2157}{10.1002/per.2157}
}
\keyword{datasets}
