% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemNetCleaner.R
\docType{package}
\name{SemNetCleaner-package}
\alias{SemNetCleaner}
\alias{SemNetCleaner-package}
\title{SemNetCleaner--package}
\description{
Implements several functions that automatize the cleaning, removal of plurals and
continuous strings, making the data binary. Also converges and finalizes linguistic
data for semantic network analysis. Uses the \code{\link{SemNetDictionaries}} package to make
the cleaning process more accurate, efficient, and reproducible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlexChristensen/SemNetCleaner}
  \item Report bugs at \url{https://github.com/AlexChristensen/SemNetCleaner/issues}
}

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
