% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemNetCleaner.R
\docType{package}
\name{SemNetCleaner-package}
\alias{SemNetCleaner}
\alias{SemNetCleaner-package}
\title{SemNetCleaner--package}
\description{
Implements several functions that automates the cleaning and
spell-checking of text data. Also converges, finalizes, removes plurals and
continuous strings, and puts text data in binary format for semantic network analysis.
Uses the \code{\link{SemNetDictionaries}} package to make
the cleaning process more accurate, efficient, and reproducible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlexChristensen/SemNetCleaner}
  \item Report bugs at \url{https://github.com/AlexChristensen/SemNetCleaner/issues}
}

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
