% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singularize.R
\name{singularize}
\alias{singularize}
\title{Converts Words to their Singular Form}
\usage{
singularize(word)
}
\arguments{
\item{word}{A word}
}
\value{
Returns the word in singular form
}
\description{
A function to de-pluralize '-s' and '-es' words. This
function is very basic and does not have many rules. The function
simply removes '-s' or '-es' from words.
}
\examples{
singularize(trial[1,10])

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
