% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singularize.R
\name{singularize}
\alias{singularize}
\title{Singularize Function}
\usage{
singularize(word)
}
\arguments{
\item{word}{A word}
}
\value{
Returns the word in singular form
}
\description{
A function to de-pluralize '-s' and '-es' words
}
\examples{

singularize(trial[1,10])

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
