% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell.opt.R
\name{spell.opt}
\alias{spell.opt}
\title{Spelling Option}
\usage{
spell.opt(word)
}
\arguments{
\item{word}{Word to check spelling options}
}
\value{
Returns a vector of all possible spelling combinations from
several spell-checking options
}
\description{
A wrapper function for spell-checking
(combines all spell-checking wrapper functions)
}
\examples{

spell.opt("bombay")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
