% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting.letter.R
\name{starting.letter}
\alias{starting.letter}
\title{Starting Letter}
\usage{
starting.letter(word)
}
\arguments{
\item{word}{Character. A single word}
}
\value{
First letter in the string
}
\description{
A wrapper function designed to produce the first
letter that appears in a word, regardless of leading characters
}
\examples{
#first letter is "w"
starting.letter("..walrus")

#first letter is "r"
starting.letter("5rat")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
